refactor: remove unused validation methods and consolidate type validation into a new TypeValidator class
