+++ app/core/services/vacancy/vacancy_get_service.py
@@ -34,7 +34,3 @@ class VacancyGetService(GetService):
     ) -> Optional[List[VacancyOutputSchema]]:
         vacancies = await self.get_all_records(filters)
         return [VacancyOutputSchema.validate(vacancy) for vacancy in vacancies]
-
-    async def get_by_name(self, name: str) -> Optional[VacancyOutputSchema]:
-        vacancy = self._get_repository.get_vacancy_by_name(name)
-        return VacancyOutputSchema.validate(vacancy)
